@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -26,10 +26,9 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-  splashscreen: any
-  font_awesome_flutter: ^8.0.1
-  http: ^0.12.0+4
+  cupertino_icons: ^1.0.4
+  font_awesome_flutter: ^9.2.0
+  http: ^0.13.4
   fluro: "^1.1.0"
 
 
