@@ -11,19 +11,20 @@ class _PerfilScreenState extends State<PerfilScreen> {
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
-      onTap: (){
+      onTap: () {
         FocusScope.of(context).unfocus();
       },
       child: Scaffold(
         drawer: CustomDrawer(),
         appBar: AppBar(
           backgroundColor: Color.fromARGB(255, 231, 172, 194),
-          title: Text("Perfil", style:
-          TextStyle(
-            color:  Colors.white,
-            fontWeight: FontWeight.w600,
-            fontSize: 26.0,
-          ),
+          title: Text(
+            "Perfil",
+            style: TextStyle(
+              color: Colors.white,
+              fontWeight: FontWeight.w600,
+              fontSize: 26.0,
+            ),
           ),
           centerTitle: true,
         ),
@@ -31,7 +32,11 @@ class _PerfilScreenState extends State<PerfilScreen> {
           child: Column(
             children: [
               Container(
-                child:Image.asset("assets/pessoa.png", height: 100.0, width: 100.0,),
+                child: Image.asset(
+                  "assets/pessoa.png",
+                  height: 100.0,
+                  width: 100.0,
+                ),
               ),
               Text(
                 "Escreva Algum Comentario: ",
@@ -42,46 +47,47 @@ class _PerfilScreenState extends State<PerfilScreen> {
               ),
               Container(
                 margin: EdgeInsets.all(14.0),
-                decoration:
-                 BoxDecoration(
-                  borderRadius:BorderRadius.circular(10.0),
+                decoration: BoxDecoration(
+                  borderRadius: BorderRadius.circular(10.0),
                   color: Colors.white,
-
                   boxShadow: [
-                     BoxShadow(
+                    BoxShadow(
                         color: Colors.black,
                         blurRadius: 3.0,
-                        offset:  Offset(1.0, 1.0))
+                        offset: Offset(1.0, 1.0))
                   ],
                 ),
                 child: TextFormField(
                   maxLines: 6,
                   maxLength: 280,
                 ),
-
               ),
-              FlatButton(
-                color: Color.fromARGB(255, 231, 172, 194),
-                child: Text("Enviar",
+              TextButton(
+                style: TextButton.styleFrom(
+                  backgroundColor: Color.fromARGB(255, 231, 172, 194),
+                  shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(20.0),
+                  ),
+                  padding: EdgeInsets.symmetric(
+                    vertical: 12.0,
+                    horizontal: 100.0,
+                  ),
+                ),
+                child: Text(
+                  "Enviar",
                   style: TextStyle(
                     color: Colors.white,
                     fontSize: 16.0,
                   ),
                 ),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(20.0),
-                ),
-                padding: EdgeInsets.symmetric(
-                  vertical: 12.0, horizontal: 100.0,),
                 onPressed: () {
-                  Navigator.of(context).push(MaterialPageRoute(
-                      builder: (context) => HomeScreen()));
+                  Navigator.of(context).push(
+                      MaterialPageRoute(builder: (context) => HomeScreen()));
                 },
               ),
             ],
           ),
         ),
-
       ),
     );
   }
