@@ -9,7 +9,6 @@ class SettingScreen extends StatefulWidget {
 }
 
 class _SettingScreenState extends State<SettingScreen> {
-
   final _scaffoldKey = GlobalKey<ScaffoldState>();
 
   final _nomeController = TextEditingController();
@@ -31,8 +30,10 @@ class _SettingScreenState extends State<SettingScreen> {
 
   @override
   Widget build(BuildContext context) {
-    return  GestureDetector(
-      onTap: (){FocusScope.of(context).unfocus();},
+    return GestureDetector(
+      onTap: () {
+        FocusScope.of(context).unfocus();
+      },
       child: Scaffold(
         key: _scaffoldKey,
         drawer: CustomDrawer(),
@@ -42,7 +43,7 @@ class _SettingScreenState extends State<SettingScreen> {
           actions: <Widget>[
             IconButton(
               icon: Icon(Icons.clear),
-              onPressed: (){
+              onPressed: () {
                 _nomeController.clear();
                 _emailController.clear();
                 _senhaController.clear();
@@ -56,12 +57,15 @@ class _SettingScreenState extends State<SettingScreen> {
             ListView(
               children: <Widget>[
                 Padding(
-                  padding: EdgeInsets.only(left:16.0, right: 16.0, top: 40.0,),
+                  padding: EdgeInsets.only(
+                    left: 16.0,
+                    right: 16.0,
+                    top: 40.0,
+                  ),
                   child: Column(
                     crossAxisAlignment: CrossAxisAlignment.stretch,
                     mainAxisAlignment: MainAxisAlignment.center,
                     children: <Widget>[
-
                       TextField(
                         decoration: InputDecoration(
                           icon: Icon(
@@ -69,21 +73,19 @@ class _SettingScreenState extends State<SettingScreen> {
                             color: Colors.black,
                           ),
                           hintText: "Insira Novo Nome",
-                          hintStyle: TextStyle(
-                              fontSize: 16.0),
+                          hintStyle: TextStyle(fontSize: 16.0),
                         ),
                         controller: _nomeController,
-                        onSubmitted: (_){
+                        onSubmitted: (_) {
                           FocusScope.of(context).requestFocus(_focusSenha);
                         },
                       ),
-
-                      const SizedBox(height: 20.0,),
-
+                      const SizedBox(
+                        height: 20.0,
+                      ),
                       TextField(
                         onSubmitted: (_) {
-                          FocusScope.of(context).requestFocus(
-                              _focusSenha);
+                          FocusScope.of(context).requestFocus(_focusSenha);
                         },
                         controller: _emailController,
                         decoration: InputDecoration(
@@ -92,13 +94,12 @@ class _SettingScreenState extends State<SettingScreen> {
                             color: Colors.black,
                           ),
                           hintText: "Insira Novo Email",
-                          hintStyle: TextStyle(
-                              fontSize: 16.0),
+                          hintStyle: TextStyle(fontSize: 16.0),
                         ),
                       ),
-
-                      const SizedBox(height: 20.0,),
-
+                      const SizedBox(
+                        height: 20.0,
+                      ),
                       TextField(
                         onSubmitted: (_) {
                           FocusScope.of(context).unfocus();
@@ -112,39 +113,38 @@ class _SettingScreenState extends State<SettingScreen> {
                             color: Colors.black,
                           ),
                           hintText: "Insira Nova Senha",
-                          hintStyle: TextStyle(
-                              fontSize: 16.0),
+                          hintStyle: TextStyle(fontSize: 16.0),
                         ),
                       ),
-
-                      const SizedBox(height: 20.0,),
-
-                      FlatButton(
-                        child: Text("Atualizar",
+                      const SizedBox(
+                        height: 20.0,
+                      ),
+                      TextButton(
+                        child: Text(
+                          "Atualizar",
                           style: TextStyle(
                               color: Colors.black,
                               fontSize: 16.0,
                               fontFamily: "WorkSansSemiBold"),
                         ),
-                        shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(20.0),
+                        style: TextButton.styleFrom(
+                          shape: RoundedRectangleBorder(
+                            borderRadius: BorderRadius.circular(20.0),
+                          ),
+                          padding: EdgeInsets.only(top: 14.0, bottom: 14.0),
                         ),
-                        padding: EdgeInsets.only(top: 14.0, bottom: 14.0),
-                        onPressed: (){
-                          if(_emailController.text.isNotEmpty &&
+                        onPressed: () {
+                          if (_emailController.text.isNotEmpty &&
                               _senhaController.text.isNotEmpty &&
-                              _nomeController.text.isNotEmpty){
+                              _nomeController.text.isNotEmpty) {
                             _onSuccess();
-                          }
-                          else  if (_emailController.text.isEmpty &&
+                          } else if (_emailController.text.isEmpty &&
                               _senhaController.text.isEmpty &&
-                              _nomeController.text.isEmpty){
+                              _nomeController.text.isEmpty) {
                             _onFail();
                           }
                         },
-
                       ),
-
                     ],
                   ),
                 ),
@@ -155,18 +155,27 @@ class _SettingScreenState extends State<SettingScreen> {
       ),
     );
   }
-  void _onSuccess(){
-    _scaffoldKey.currentState.showSnackBar( SnackBar(
-      content: Text("Informaçôes atualizadas com sucesso",),
-      duration: Duration(seconds: 2,),
+
+  void _onSuccess() {
+    _scaffoldKey.currentState!.showSnackBar(SnackBar(
+      content: Text(
+        "Informaçôes atualizadas com sucesso",
+      ),
+      duration: Duration(
+        seconds: 2,
+      ),
       backgroundColor: Color.fromARGB(255, 4, 125, 141),
     ));
   }
 
-  void _onFail(){
-    _scaffoldKey.currentState.showSnackBar( SnackBar(
-      content: Text("Falha ao atualizar as informações usuário ",),
-      duration: Duration(seconds: 2,),
+  void _onFail() {
+    _scaffoldKey.currentState!.showSnackBar(SnackBar(
+      content: Text(
+        "Falha ao atualizar as informações usuário ",
+      ),
+      duration: Duration(
+        seconds: 2,
+      ),
       backgroundColor: Color.fromARGB(255, 4, 125, 141),
     ));
   }
