@@ -5,8 +5,7 @@ class GoogleButton extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
-      margin: const EdgeInsets.only(
-          left: 30.0, right: 30.0, top: 20.0),
+      margin: const EdgeInsets.only(left: 30.0, right: 30.0, top: 20.0),
       child: Row(
         children: <Widget>[
           Expanded(
@@ -14,31 +13,32 @@ class GoogleButton extends StatelessWidget {
               margin: EdgeInsets.only(right: 8.0),
               alignment: Alignment.center,
               child: Row(
-
                 children: <Widget>[
                   Expanded(
-                    child: FlatButton(
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius
-                            .circular(30.0),
-                      ),
-                      color: Color(0Xffdb3236),
+                    child: TextButton(
                       onPressed: () => {},
+                      style: TextButton.styleFrom(
+                        shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(30.0),
+                        ),
+                        backgroundColor: Color(0Xffdb3236),
+                      ),
                       child: Container(
                         child: Row(
-                          mainAxisAlignment: MainAxisAlignment
-                              .center,
+                          mainAxisAlignment: MainAxisAlignment.center,
                           children: <Widget>[
                             Expanded(
-                              child: FlatButton(
+                              child: TextButton(
                                 onPressed: () => {},
-                                padding: EdgeInsets
-                                    .only(
-                                  top: 20.0,
-                                  bottom: 20.0,
+                                style: TextButton.styleFrom(
+                                  padding: EdgeInsets.only(
+                                    top: 20.0,
+                                    bottom: 20.0,
+                                  ),
                                 ),
-                                child:Row(
-                                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                                child: Row(
+                                  mainAxisAlignment:
+                                      MainAxisAlignment.spaceEvenly,
                                   children: [
                                     Icon(
                                       FontAwesomeIcons.google,
@@ -46,17 +46,13 @@ class GoogleButton extends StatelessWidget {
                                     ),
                                     Text(
                                       "GOOGLE",
-                                      textAlign: TextAlign
-                                          .center,
+                                      textAlign: TextAlign.center,
                                       style: TextStyle(
-                                          color: Colors
-                                              .white,
-                                          fontWeight: FontWeight
-                                              .bold),
+                                          color: Colors.white,
+                                          fontWeight: FontWeight.bold),
                                     ),
                                   ],
                                 ),
-
                               ),
                             ),
                           ],
