@@ -1,84 +1,35 @@
 class Api {
   List<News> news;
 
-  Api({this.news});
+  Api({required this.news});
 
-  Api.fromJson(Map<String, dynamic> json) {
-    if (json['news'] != null) {
-      news = new List<News>();
-      json['news'].forEach((v) {
-        news.add(new News.fromJson(v));
-      });
-    }
+
   }
 
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    if (this.news != null) {
-      data['news'] = this.news.map((v) => v.toJson()).toList();
-    }
-    return data;
-  }
-}
 
 class News {
-  User user;
-  Message message;
+  final User user;
+  final Message message;
 
-  News({this.user, this.message});
+  News({required this.user, required this.message});
 
-  News.fromJson(Map<String, dynamic> json) {
-    user = json['user'] != null ? new User.fromJson(json['user']) : null;
-    message =
-    json['message'] != null ? new Message.fromJson(json['message']) : null;
-  }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    if (this.user != null) {
-      data['user'] = this.user.toJson();
-    }
-    if (this.message != null) {
-      data['message'] = this.message.toJson();
-    }
-    return data;
-  }
 }
 
 class User {
-  String name;
-  String profilePicture;
+  final String name;
+  final String profilePicture;
 
-  User({this.name, this.profilePicture});
+  User({required this.name, required this.profilePicture});
 
-  User.fromJson(Map<String, dynamic> json) {
-    name = json['name'];
-    profilePicture = json['profile_picture'];
-  }
 
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    data['name'] = this.name;
-    data['profile_picture'] = this.profilePicture;
-    return data;
-  }
 }
 
 class Message {
-  String content;
-  String createdAt;
+  final String content;
+  final String createdAt;
 
-  Message({this.content, this.createdAt});
+  Message({required this.content, required this.createdAt});
 
-  Message.fromJson(Map<String, dynamic> json) {
-    content = json['content'];
-    createdAt = json['created_at'];
-  }
 
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    data['content'] = this.content;
-    data['created_at'] = this.createdAt;
-    return data;
-  }
+
 }
